[Bronze II] Title: Hashing, Time: 164 ms, Memory: 17784 KB, Score: 50 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -9,12 +9,12 @@ public class Main
 	    sc.nextLine();
 	    String str = sc.nextLine();
 	    
-	    int sum = 0;
+	    long sum = 0;
 	    for(int j=0; j<str.length(); j++) {
-	        int t = str.charAt(j) - 'a' + 1;
+	        long t = str.charAt(j) - 'a' + 1;
 	        
 	        
-	        sum = sum + t * (int)Math.pow(31, j);
+	        sum = sum + t * (long)Math.pow(31, j);
 	        
 	    }
 	    sum = sum%1234567891;
